==> 2-Feature_Engineering.ipynb <==
# Earnings Call Project: Feature Engineering
<br>
CIS 831 Deep Learning – Term Project<br>
Kansas State University
<br><br>
James Chapman<br>
John Woods<br>
Nathan Diehl<br>
<br>

This notebook featurizes the text and audio data from the earnings calls. Each earnings calls data comes pre-processed such that each sentence of the call corresponds to 1 line-of-text and 1 MP3 audio file. The transcript text is processed with the Glove–300 pre-trained word embedding, and the audio files are processed with Praat using parselmouth.
- Text (Glove–300) [Glove Download](https://nlp.stanford.edu/projects/glove/)
- Audio (Praat) [Parselmouth](https://parselmouth.readthedocs.io/en/stable/)

The data from this notebook is stored in the "data/data_prep" directory as the following CSVs.
- glove_features
- MAEC_glove_features
- praat_features
- MAEC_praat_features

import sys
IN_COLAB = "google.colab" in sys.modules

if IN_COLAB:
    !pip install parselmouth
    from google.colab import drive
    drive.mount('/content/gdrive')
    %cd gdrive/My Drive/831

import pandas as pd
import numpy as np
import parselmouth
import re
import os
from datetime import datetime
from tqdm import tqdm

MAEC_dir = 'data/MAEC/MAEC_Dataset' # https://github.com/Earnings-Call-Dataset/MAEC-A-Multimodal-Aligned-Earnings-Conference-Call-Dataset-for-Financial-Risk-Prediction

############# too big for GitHub ########################
############# stored on local disk ######################
original_data_dir = r"D:\original_dataset" # https://github.com/GeminiLn/EarningsCall_Dataset 
MAEC_audio_dir = r"D:\MAEC_audio" 
# there is a link for the audio data in the MAEC GitHub, but it does not work
# I emailed the authors, and they send another link.
# There is like a half-million files, but only 19 GB
# https://drive.google.com/file/d/1m1GRCHgKn9Vz9IFMC_SpCog6uP3-gFgY/view?usp=drive_link 

# Loop through the directory, each folder represents an earnings conference call; the folders are named as "CompanyName_Date".
filename_data = []
for filename in os.listdir(original_data_dir):
    company_name, date_str = filename.rsplit('_', 1)
    date_str = date_str.split('.')[0] 
    date = datetime.strptime(date_str, "%Y%m%d").strftime("%Y-%m-%d")
    filename_data.append([company_name, date])
filename_data = pd.DataFrame(filename_data, columns=["Company", "Date"])
company_ticker = pd.read_csv('data/data_prep/company_ticker.csv')
filename_data = filename_data.merge(company_ticker, on="Company", how="left")

# Loop through the directory, each folder represents an earnings conference call; the folders are named as "Date_CompanyName".
MAEC_filename_data = []
for filename in os.listdir(MAEC_dir):
    date_str, ticker = filename.rsplit('_', 1)
    date_str = date_str.split('.')[0] 
    date = datetime.strptime(date_str, "%Y%m%d").strftime("%Y-%m-%d")
    MAEC_filename_data.append([ticker, date])
MAEC_filename_data = pd.DataFrame(MAEC_filename_data, columns=["Ticker", "Date"])

Here, we loop through the original dataset (text & audio) and show each sentence should have one line of text & one audio file.

def each_row(row):
    Company = row['Company']
    Date = row['Date'].replace('-', '') 

    text_path = f"D:/original_dataset/{Company}_{Date}/TextSequence.txt"
    with open(text_path, 'r', encoding='utf-8', errors='replace') as file:
        for i, line in enumerate(file, start= 1):
            #I'll do something with this 
            text_length = i

    audio_length = 0
    audio_dir = f"D:/original_dataset/{Company}_{Date}/CEO"
    if os.path.exists(audio_dir):
        for i, audio_file in enumerate(os.listdir(audio_dir), start= 1):
            #I'll do something with this
            if audio_file.lower().endswith('.mp3'):
                audio_length += 1
            
    return text_length, audio_length

filename_data[['text_length', 'audio_length']] = filename_data.apply(each_row, axis=1, result_type='expand')
print('should be zero-', len(filename_data[filename_data['text_length']!= filename_data['audio_length']] ))
filename_data.info(verbose=True)

def each_row(row):
    Ticker = row['Ticker']
    Date = row['Date'].replace('-', '') 

    text_path = f"D:/MAEC_audio/{Date}_{Ticker}/text.txt"
    with open(text_path, 'r', encoding='utf-8', errors='replace') as file:
        for i, line in enumerate(file, start= 1):
            #I'll do something with this 
            text_length = i
            
    audio_length = 0
    audio_dir = f"D:/MAEC_audio/{Date}_{Ticker}"
    if os.path.exists(audio_dir):
        for i, audio_file in enumerate(os.listdir(audio_dir), start= 1):
            #I'll do something with this
            if audio_file.lower().endswith('.mp3'):
                audio_length += 1
            
    return text_length, audio_length

MAEC_filename_data[['text_length', 'audio_length']] = MAEC_filename_data.apply(each_row, axis=1, result_type='expand')
print('should be zero-', len(MAEC_filename_data[MAEC_filename_data['text_length']!= MAEC_filename_data['audio_length']] ))
MAEC_filename_data.info(verbose=True)

# Glove 300 features from meeting transcript text files

Glove file can be found at https://nlp.stanford.edu/projects/glove/

glove_file_path = r"C:\Users\James\OneDrive\Kansas State University\CIS 831\Project\glove.6B\glove.6B.300d.txt" 
glove_model = {}
with open(glove_file_path, 'r', encoding='utf-8') as f:
    for line in f:
        parts = line.split()
        word = parts[0]
        vector = np.array(parts[1:], dtype='float32')
        glove_model[word] = vector
print(f"Loaded {len(glove_model)} words from GloVe.")

contractions = {
    "i'll": "i will", "can't": "cannot", "won't": "will not",
    "i'm": "i am", "he's": "he is", "she's": "she is",
    "they're": "they are", "it's": "it is", "let's": "let us",
    "don't": "do not", "didn't": "did not", "doesn't": "does not",
    "you're": "you are", "we're": "we are", "they've": "they have",
    "there's": "there is", "who's": "who is", "what's": "what is",
    "i'd": "i would", "he'd": "he would", "she'd": "she would",
    "they'd": "they would", "we'd": "we would", "you'd": "you would",
    "i've": "i have", "we've": "we have", "you've": "you have",
    "shouldn't": "should not", "wouldn't": "would not", "couldn't": "could not",
    "isn't": "is not", "aren't": "are not", "wasn't": "was not",
    "weren't": "were not", "hasn't": "has not", "haven't": "have not",
    "hadn't": "had not", "mustn't": "must not", "mightn't": "might not",
    "needn't": "need not", "shan't": "shall not", "would've": "would have",
    "should've": "should have", "could've": "could have", "might've": "might have",
    "must've": "must have", "let's": "let us", "he'll": "he will",
    "she'll": "she will", "they'll": "they will", "we'll": "we will",
    "you'll": "you will", "that'll": "that will", "there'll": "there will",
    "it'll": "it will", "who'll": "who will", "how's": "how is",
    "why's": "why is", "where's": "where is", "when's": "when is",
    "ain't": "is not", "couldn't've": "could not have", "shouldn't've": "should not have",
    "wouldn't've": "would not have", "oughtn't": "ought not", "she'd've": "she would have",
    "he'd've": "he would have", "they'd've": "they would have", "we'd've": "we would have",
    "you'd've": "you would have", "i'd've": "i would have", "it'd've": "it would have",
    "y'all": "you all", "how'd": "how did", "what'd": "what did",
    "who'd": "who did", "where'd": "where did", "when'd": "when did",
    "why'd": "why did", "how'll": "how will", "what'll": "what will",
    "who'll": "who will", "where'll": "where will", "when'll": "when will",
    "why'll": "why will"
}

# Returns 300 dimensional vector (averaged of word vectors). Is there a better way?
def get_glove(sentence):
    # clean up each sentence
    sentence = sentence.lower()
    for contraction, full_form in contractions.items():
        sentence = sentence.replace(contraction, full_form)
    sentence = re.sub(r'[^a-z0-9\s]', ' ', sentence)

    words = sentence.split()
    word_embeddings = []
    for word in words:
        if word in glove_model:
            word_embeddings.append(glove_model[word])
    if not word_embeddings:
        print(sentence)
        return [np.zeros(300)], True  # Return a zero vector if no words match
    return [np.mean(word_embeddings, axis=0)], False

glove_features = pd.DataFrame()
bad_glove = []
for Company, Date in tqdm(filename_data[['Company', 'Date']].values):
    Date = Date.replace('-', '')
    text_path = f"D:/original_dataset/{Company}_{Date}/TextSequence.txt"
    with open(text_path, 'r', encoding='utf-8', errors='replace') as file:
        for i, line in enumerate(file, start=1):
            sentence_embedding, sentence_has_error = get_glove(line.strip())
            if sentence_has_error: # whenever a sentence is returned as all zeros
                bad_glove.append([Company,Date,i,line.strip()])
            features_df = pd.DataFrame(sentence_embedding, columns=[f'GloVe_{j}' for j in range(300)])
            features_df['Company'] = Company
            features_df['Date'] = Date
            features_df['Sentence_num'] = i
            glove_features = pd.concat([glove_features, features_df], ignore_index=True)

print(len(bad_glove))
# printed below are the sentences without word embeddings (all zeros)

glove_features.info(verbose=True)
### save ############################################
glove_features.to_csv('data/data_prep/glove_features.csv', index=False)
#####################################################

MAEC_glove_features = pd.DataFrame()
MAEC_bad_glove = []
for Ticker, Date in tqdm(MAEC_filename_data[['Ticker', 'Date']].values):
    Date = Date.replace('-', '')
    text_path = f"D:/MAEC_audio/{Date}_{Ticker}/text.txt"
    with open(text_path, 'r', encoding='utf-8', errors='replace') as file:
        for i, line in enumerate(file, start=1):
            sentence_embedding, sentence_has_error = get_glove(line.strip())
            if sentence_has_error:  # whenever a sentence is returned as all zeros
                MAEC_bad_glove.append([Company,Date,i,line.strip()])
            features_df = pd.DataFrame(sentence_embedding, columns=[f'GloVe_{j}' for j in range(300)])
            features_df['Ticker'] = Ticker
            features_df['Date'] = Date
            features_df['Sentence_num'] = i
            MAEC_glove_features = pd.concat([MAEC_glove_features, features_df], ignore_index=True)

print(len(MAEC_bad_glove))
# printed below are the sentences without word embeddings (all zeros)

MAEC_glove_features.info(verbose=True)
### save ############################################
MAEC_glove_features.to_csv('data/data_prep/MAEC_glove_features.csv', index=False)
#####################################################

# Praat features from MP3 audio files

Using parselmouth, "Voice report" produces this following string.

sound = parselmouth.Sound(r"D:\MAEC_audio\20150225_LMAT\LMAT_20150225_f000002100.mp3")
pitch = sound.to_pitch()
pulses = parselmouth.praat.call([sound, pitch], "To PointProcess (cc)")
voice_report_str = parselmouth.praat.call([sound, pitch, pulses], "Voice report", 0.0, 0.0, 75, 600, 1.3, 1.6, 0.03, 0.45)
print(voice_report_str) 

# Returns dictionary of 29 features
def get_praat(audio_dir,audio_file):
    sound = parselmouth.Sound(os.path.join(audio_dir,audio_file))
    pitch = sound.to_pitch()
    pulses = parselmouth.praat.call([sound, pitch], "To PointProcess (cc)")
    voice_report_str = parselmouth.praat.call([sound, pitch, pulses], "Voice report", 0.0, 0.0, 75, 600, 1.3, 1.6, 0.03, 0.45)
    
    # not in voice report
    audio_length = sound.get_total_duration()
    try:
        intensity = sound.to_intensity()
        mean_intensity = intensity.values.mean()
        min_intensity = intensity.values.min()
        max_intensity = intensity.values.max()
    except: 
        mean_intensity = '--undefined--'
        min_intensity = '--undefined--'
        max_intensity = '--undefined--'
        audio_length = '--undefined--'

    mean_pitch = re.search(r'Mean pitch: ([\d.]+|--undefined--)', voice_report_str).group(1)
    stdev_pitch = re.search(r'Standard deviation: ([\d.]+|--undefined--)', voice_report_str).group(1)
    min_pitch = re.search(r'Minimum pitch: ([\d.]+|--undefined--)', voice_report_str).group(1)
    max_pitch = re.search(r'Maximum pitch: ([\d.]+|--undefined--)', voice_report_str).group(1)
    num_pulses = re.search(r'Number of pulses: (\d+|--undefined--)', voice_report_str).group(1)
    num_periods = re.search(r'Number of periods: (\d+|--undefined--)', voice_report_str).group(1)
    mean_period = re.search(r'Mean period: ([\d.E+-]+|--undefined--)', voice_report_str).group(1)
    stdev_period = re.search(r'Standard deviation of period: ([\d.E+-]+|--undefined--)', voice_report_str).group(1)
    fraction_unvoiced = re.search(r'Fraction of locally unvoiced frames: ([\d.]+|--undefined--)%?', voice_report_str).group(1)
    num_voice_breaks = re.search(r'Number of voice breaks: (\d+|--undefined--)', voice_report_str).group(1)
    degree_of_voice_breaks = re.search(r'Degree of voice breaks: ([\d.]+|--undefined--)', voice_report_str).group(1)
    jitter_local = re.search(r'Jitter \(local\): ([\d.]+|--undefined--)%?', voice_report_str).group(1)
    jitter_local_abs = re.search(r'Jitter \(local, absolute\): ([\d.E+-]+|--undefined--)', voice_report_str).group(1)
    jitter_rap = re.search(r'Jitter \(rap\): ([\d.]+|--undefined--)%?', voice_report_str).group(1)
    jitter_ppq5 = re.search(r'Jitter \(ppq5\): ([\d.]+|--undefined--)%?', voice_report_str).group(1)
    jitter_ddp = re.search(r'Jitter \(ddp\): ([\d.]+|--undefined--)%?', voice_report_str).group(1)
    shimmer_local = re.search(r'Shimmer \(local\): ([\d.]+|--undefined--)%?', voice_report_str).group(1)
    shimmer_local_dB = re.search(r'Shimmer \(local, dB\): ([\d.]+|--undefined--) dB', voice_report_str).group(1)
    shimmer_apq3 = re.search(r'Shimmer \(apq3\): ([\d.]+|--undefined--)%?', voice_report_str).group(1)
    shimmer_apq5 = re.search(r'Shimmer \(apq5\): ([\d.]+|--undefined--)%?', voice_report_str).group(1)
    shimmer_apq11 = re.search(r'Shimmer \(apq11\): ([\d.]+|--undefined--)', voice_report_str).group(1)
    shimmer_dda = re.search(r'Shimmer \(dda\): ([\d.]+|--undefined--)%?', voice_report_str).group(1)
    autocorrelation = re.search(r'Mean autocorrelation: ([\d.]+|--undefined--)', voice_report_str).group(1)
    nhr = re.search(r'Mean noise-to-harmonics ratio: ([\d.]+|--undefined--)', voice_report_str).group(1)
    try:
        hnr = re.search(r'Mean harmonics-to-noise ratio: ([\d.]+|--undefined--) dB', voice_report_str).group(1)
    except: 
        hnr = '--undefined--'

    features = {
        'Mean pitch': mean_pitch,
        'Standard deviation': stdev_pitch,
        'Minimum pitch': min_pitch,
        'Maximum pitch': max_pitch,
        'Number of pulses': num_pulses,
        'Number of periods': num_periods,
        'Mean period': mean_period,
        'Mean intensity': mean_intensity,
        'Minimum intensity': min_intensity,
        'Maximum intensity': max_intensity,
        'Standard deviation of period': stdev_period,
        'Fraction of unvoiced': fraction_unvoiced,  
        'Number of voice breaks': num_voice_breaks, 
        'Degree of voice breaks': degree_of_voice_breaks,  
        'Jitter local': jitter_local,
        'Jitter local absolute': jitter_local_abs,
        'Jitter rap': jitter_rap,
        'Jitter ppq5': jitter_ppq5,
        'Jitter ddp': jitter_ddp,
        'Shimmer local': shimmer_local,
        'Shimmer local dB': shimmer_local_dB,
        'Shimmer apq3': shimmer_apq3,
        'Shimmer apq5': shimmer_apq5,
        'Shimmer apq11': shimmer_apq11,
        'Shimmer dda': shimmer_dda,
        'Mean autocorrelation': autocorrelation,
        'Mean NHR': nhr,
        'Mean HNR': hnr,
        'Audio Length': audio_length
    }
    return features

# Collect whenever there is an error
bad_praat = [] # Company,Date,i,audio_file,voice_report_str,e
# set all features to undefined
undefined_features = {'Mean pitch': '--undefined--', 'Standard deviation':'--undefined--', 'Minimum pitch': '--undefined--', 'Maximum pitch': '--undefined--', 'Number of pulses': '--undefined--', 'Number of periods': '--undefined--', 'Mean period':  '--undefined--', 'Mean intensity': '--undefined--', 'Minimum intensity':  '--undefined--', 'Maximum intensity':  '--undefined--', 'Standard deviation of period': '--undefined--', 'Fraction of unvoiced':'--undefined--', 'Number of voice breaks':  '--undefined--', 'Degree of voice breaks':  '--undefined--', 'Jitter local': '--undefined--', 'Jitter local absolute': '--undefined--', 'Jitter rap':  '--undefined--', 'Jitter ppq5':  '--undefined--', 'Jitter ddp': '--undefined--', 'Shimmer local':'--undefined--', 'Shimmer local dB': '--undefined--', 'Shimmer apq3': '--undefined--', 'Shimmer apq5': '--undefined--', 'Shimmer apq11': '--undefined--', 'Shimmer dda': '--undefined--', 'Mean autocorrelation':'--undefined--', 'Mean NHR': '--undefined--', 'Mean HNR': '--undefined--', 'Audio Length': '--undefined--'}

praat_features = pd.DataFrame()
for Company,Date in tqdm(filename_data[['Company','Date']].values):
    Date = Date.replace('-', '') 
    audio_dir = f"D:/original_dataset/{Company}_{Date}/CEO"
    if os.path.exists(audio_dir):
        for i, audio_file in enumerate(os.listdir(audio_dir), start= 1):
            try:
                # skip files that are not MP3 audio
                if audio_file.lower().endswith('.mp3'):
                    features = get_praat(audio_dir,audio_file)
                    features_df = pd.DataFrame([features])
                    features_df['Company'] = Company
                    features_df['Date'] = Date
                    features_df['Sentence_num'] = i
                    features_df['audio_file'] = audio_file
                    praat_features = pd.concat([praat_features, features_df], ignore_index=True)
            except KeyboardInterrupt: break
            except Exception as e: 
                bad_praat.append([Company,Date,i,audio_file,voice_report_str,e])
                # set all features to undefined
                features_df = pd.DataFrame([undefined_features])
                features_df['Company'] = Company
                features_df['Date'] = Date
                features_df['Sentence_num'] = i
                features_df['audio_file'] = audio_file
                praat_features = pd.concat([praat_features, features_df], ignore_index=True)

praat_features.info(verbose=True)
### save ############################################
praat_features.to_csv('data/data_prep/praat_features.csv', index=False)
#####################################################

print(len(bad_praat))
bad_praat

bad_MAEC_praat = []
MAEC_praat_features = pd.DataFrame()
for Ticker,Date in tqdm(MAEC_filename_data[['Ticker','Date']].values):
    Date = Date.replace('-', '') 
    audio_dir = f"D:/MAEC_audio/{Date}_{Ticker}"
    if os.path.exists(audio_dir):
        for i, audio_file in enumerate(os.listdir(audio_dir), start= 1):
            try:
                # skip files that are not MP3 audio
                if audio_file.lower().endswith('.mp3'):
                    features = get_praat(audio_dir,audio_file)
                    features_df = pd.DataFrame([features])
                    features_df['Ticker'] = Ticker
                    features_df['Date'] = Date
                    features_df['Sentence_num'] = i
                    features_df['audio_file'] = audio_file
                    MAEC_praat_features = pd.concat([MAEC_praat_features, features_df], ignore_index=True)
            except KeyboardInterrupt: break
            except Exception as e: 
                bad_praat.append([Ticker,Date,i,audio_file,voice_report_str,e])
                # set all features to undefined
                features_df = pd.DataFrame([undefined_features])
                features_df['Ticker'] = Ticker
                features_df['Date'] = Date
                features_df['Sentence_num'] = i
                features_df['audio_file'] = audio_file
                MAEC_praat_features = pd.concat([MAEC_praat_features, features_df], ignore_index=True)


MAEC_praat_features.info(verbose=True)
### save ############################################
MAEC_praat_features.to_csv('data/data_prep/MAEC_praat_features.csv', index=False)
#####################################################

print(len(bad_MAEC_praat))
bad_MAEC_praat

print(bad_MAEC_praat)
