==> 2c-Feature_Engineering_emotion2vec.ipynb <==
# Earnings Call Project: emotion2vec
<br>
CIS 831 Deep Learning – Term Project<br>
Kansas State University
<br><br>
James Chapman<br>
John Woods<br>
Nathan Diehl<br>
<br>

### This notebook featurizes the AUDIO data from the earnings calls with emotion2vec.

emotion2vec documentation can be found at https://github.com/ddlBoJack/emotion2vec
and
https://huggingface.co/emotion2vec

The data from this notebook is stored in the "data/data_prep" directory as the following CSVs.

* emotion2vec
* MAEC_emotion2vec


import pandas as pd
import numpy as np
import os
from datetime import datetime
from tqdm import tqdm
import torch
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

MAEC_dir = 'data/MAEC/MAEC_Dataset' # https://github.com/Earnings-Call-Dataset/MAEC-A-Multimodal-Aligned-Earnings-Conference-Call-Dataset-for-Financial-Risk-Prediction

############# too big for GitHub ########################
############# stored on local disk ######################
original_data_dir = r"D:\original_dataset" # https://github.com/GeminiLn/EarningsCall_Dataset 
MAEC_audio_dir = r"D:\MAEC_audio" 
# there is a link for the audio data in the MAEC GitHub, but it does not work
# I emailed the authors, and they send another link.
# There is like a half-million files, but only 19 GB
# https://drive.google.com/file/d/1m1GRCHgKn9Vz9IFMC_SpCog6uP3-gFgY/view?usp=drive_link 

# Loop through the directory, each folder represents an earnings conference call; the folders are named as "CompanyName_Date".
filename_data = []
for filename in os.listdir(original_data_dir):
    company_name, date_str = filename.rsplit('_', 1)
    date_str = date_str.split('.')[0] 
    date = datetime.strptime(date_str, "%Y%m%d").strftime("%Y-%m-%d")
    filename_data.append([company_name, date])
filename_data = pd.DataFrame(filename_data, columns=["Company", "Date"])
company_ticker = pd.read_csv('data/data_prep/company_ticker.csv')
filename_data = filename_data.merge(company_ticker, on="Company", how="left")

# Loop through the directory, each folder represents an earnings conference call; the folders are named as "Date_CompanyName".
MAEC_filename_data = []
for filename in os.listdir(MAEC_dir):
    date_str, ticker = filename.rsplit('_', 1)
    date_str = date_str.split('.')[0] 
    date = datetime.strptime(date_str, "%Y%m%d").strftime("%Y-%m-%d")
    MAEC_filename_data.append([ticker, date])
MAEC_filename_data = pd.DataFrame(MAEC_filename_data, columns=["Ticker", "Date"])

from funasr import AutoModel

model = AutoModel(model="iic/emotion2vec_plus_large")

# example of using the model, and the output
# I included the features in this one
MP3_file = "D:/original_dataset/3M Company_20170425/CEO/Nicholas C. Gangestad_1_1.mp3"
results = model.generate(MP3_file, output_dir="./outputs", granularity="utterance", extract_embedding=True)
labels = results[0]['labels']
scores = results[0]['scores']
feats = results[0]['feats']

print(results)
print(labels)
print(scores)
print(len(feats))


import sys
import os

# the output of the model kept freezing my computer
# each audio file seems to print out a progress bar

# Suppress output temporarily
class SuppressOutput:
    def __enter__(self):
        self._original_stdout = sys.stdout
        self._original_stderr = sys.stderr
        sys.stdout = open(os.devnull, 'w')
        sys.stderr = open(os.devnull, 'w')

    def __exit__(self, exc_type, exc_val, exc_tb):
        sys.stdout.close()
        sys.stderr.close()
        sys.stdout = self._original_stdout
        sys.stderr = self._original_stderr




# I kept having a memory error
# this splits up each audio file into sections at most 2 minutes
# then averages the output factors
from pydub import AudioSegment
segment_length = 120000 # 2 minutes

MP3_file = "D:/MAEC_audio/20160802_CORT/CORT_20160802_f000023106.mp3"
audio = AudioSegment.from_file(MP3_file, format="mp3")
segments = [audio[i:i + segment_length] for i in range(0, len(audio), segment_length)]
print(len(segments))

all_scores = []
for i, segment in enumerate(segments):
    temp_file = f"temp_segment_{i}.mp3"
    segment.export(temp_file, format="mp3")  
    results = model.generate(temp_file, output_dir="./outputs", granularity="utterance", extract_embedding=False)
    print(results[0]['scores'])
    all_scores.append(results[0]['scores'])

scores = [sum(x) / len(all_scores) for x in zip(*all_scores)]
print(scores)
    


# Returns 9 floating-point values 
def get_emotion2vec(audio_dir,audio_file):
    MP3_file = os.path.join(audio_dir,audio_file)
    audio = AudioSegment.from_file(MP3_file, format="mp3")
    # split MP3 files into sections of 2 minutes
    segments = [audio[i:i + segment_length] for i in range(0, len(audio), segment_length)]

    all_scores = []
    for i, segment in enumerate(segments):
        temp_file = f"temp_segment_{i}.mp3"
        segment.export(temp_file, format="mp3")  
        with SuppressOutput():
            results = model.generate(temp_file, output_dir="./outputs", granularity="utterance", extract_embedding=False)
        labels = results[0]['labels']
        all_scores.append(results[0]['scores'])
        torch.cuda.empty_cache() # I don't know if this helps
    # average all of the sections into one factor of 9
    scores = [sum(x) / len(all_scores) for x in zip(*all_scores)]
    return labels, scores


bad_emotion2vec = [] # Company ,Date, i, audio_dir, audio_file, labels, scores, e

emotion2vec = []
for Company,Date in filename_data[['Company','Date']].values: # each audio file of the original data set
    Date = Date.replace('-', '') 
    audio_dir = f"D:/original_dataset/{Company}_{Date}/CEO"
    if os.path.exists(audio_dir):
        for i, audio_file in enumerate(os.listdir(audio_dir), start= 1):
            try:
                # skip files that are not MP3 audio
                if audio_file.lower().endswith('.mp3'):
                    labels, scores = get_emotion2vec(audio_dir,audio_file)
                    features_row = np.concatenate([labels, scores, [Company, Date, i, audio_file]])
                    emotion2vec.append(features_row)
            except KeyboardInterrupt: break
            except Exception as e: 
                print(Company ,Date, i, audio_dir, audio_file, e)
                bad_emotion2vec.append([Company ,Date, i, audio_dir, audio_file, labels, scores, e])
                # set all features to  0.0 if there was an error
                scores = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
                features_row = np.concatenate([labels, scores, [Company, Date, i, audio_file]])
                emotion2vec.append(features_row)
                   
emotion2vec = pd.DataFrame(emotion2vec)
emotion2vec.info(verbose=True)
emotion2vec.describe()

# check to make sure all the labels (the first 9 columns) are the same
# then drop them
for column in range(9):
    print('Should be 1 ---', pd.unique(emotion2vec[column]).size)
emotion2vec = emotion2vec.iloc[:, 9:]

# rename columns
emotion2vec.columns = ['angry', 'disgusted', 'fearful', 'happy', 'neutral', 'other', 'sad', 'surprised', 'unk', 
                       'Company', 'Date', 'Sentence_num', 'audio_file']
emotion2vec = emotion2vec.drop(['audio_file'], axis=1)
emotion2vec.info(verbose=True)
### save ############################################
emotion2vec.to_csv('data/data_prep/emotion2vec.csv', index=False)
#####################################################
emotion2vec.describe()


bad_emotion2vec

# MAEC


# Collect whenever there is an error
MAEC_bad_emotion2vec = [] # Ticker , Date, i, audio_dir, audio_file, labels, scores, e

MAEC_emotion2vec = []
for Ticker,Date in tqdm(MAEC_filename_data[['Ticker','Date']].values): # each audio file of the original data set
    Date = Date.replace('-', '') 
    audio_dir = f"D:/MAEC_audio/{Date}_{Ticker}"
    if os.path.exists(audio_dir):
        for i, audio_file in enumerate(os.listdir(audio_dir), start= 1):
            try:
                # skip files that are not MP3 audio
                if audio_file.lower().endswith('.mp3'):
                    labels, scores = get_emotion2vec(audio_dir,audio_file)
                    features_row = np.concatenate([labels, scores, [Ticker, Date, i, audio_file]])
                    MAEC_emotion2vec.append(features_row)
            except KeyboardInterrupt: break
            except Exception as e: 
                print(Ticker, Date, i, e)
                MAEC_bad_emotion2vec.append([Ticker ,Date, i, audio_dir, audio_file, labels, scores, e])
                # set all features to  0.0 if there was an error
                scores = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
                features_row = np.concatenate([labels, scores, [Ticker, Date, i, audio_file]])
                MAEC_emotion2vec.append(features_row)

MAEC_emotion2vec = pd.DataFrame(MAEC_emotion2vec)
MAEC_emotion2vec.info(verbose=True)
MAEC_emotion2vec.describe()

# check to make sure all the labels (the first 9 columns) are the same
# then drop them
for column in range(9):
    print('Should be 1 ---', pd.unique(MAEC_emotion2vec[column]).size)
MAEC_emotion2vec = MAEC_emotion2vec.iloc[:, 9:]

# rename columns
MAEC_emotion2vec.columns = ['angry', 'disgusted', 'fearful', 'happy', 'neutral', 'other', 'sad', 'surprised', 'unk', 
                       'Company', 'Date', 'Sentence_num', 'audio_file']
MAEC_emotion2vec = MAEC_emotion2vec.drop(['audio_file'], axis=1)
MAEC_emotion2vec.info(verbose=True)
### save ############################################
MAEC_emotion2vec.to_csv('data/data_prep/MAEC_emotion2vec.csv', index=False)
#####################################################
MAEC_emotion2vec.describe()

MAEC_bad_emotion2vec

print(len(MAEC_bad_emotion2vec))
