==> 3-Data_Cleaning.ipynb <==
# Earnings Call Project: Data Cleaning
<br>
CIS 831 Deep Learning – Term Project<br>
Kansas State University
<br><br>
James Chapman<br>
John Woods<br>
Nathan Diehl<br>
<br>

This notebook creates data used for training/testing.
- Calculates the targets for both datasets
- Corrects Praat features for both datasets
- Combines features audio (Praat) & text (6 different) and targets
- save 9 numpy files specifically for HTML (for each audio/text pair)
    - train (features, targets, secondary_targets)
    - validation (features, targets, secondary_targets)
    - test (features, targets, secondary_targets)
- it is important to note 12 meetings were removed because we could not find stock data (Yahoo or alphadvantage)<br>
- 218 meetings were removed from MAEC <br>

The rest data from this notebook is stored in the "data" directory as the following CSVs (for each audio/text pair)
- MAEC_dataset
- MAEC_dataset

import pandas as pd
import numpy as np
import os
from datetime import datetime
from tqdm import tqdm

tqdm.pandas()

MAEC_dir = 'data/MAEC/MAEC_Dataset' # https://github.com/Earnings-Call-Dataset/MAEC-A-Multimodal-Aligned-Earnings-Conference-Call-Dataset-for-Financial-Risk-Prediction

############# too big for GitHub ########################
############# stored on local disk ######################
original_data_dir = r"D:\original_dataset" # https://github.com/GeminiLn/EarningsCall_Dataset 
MAEC_audio_dir = r"D:\MAEC_audio" 
# there is a link for the audio data in the MAEC GitHub, but it does not work
# I emailed the authors, and they send another link.
# There is like a half-million files, but only 19 GB
# https://drive.google.com/file/d/1m1GRCHgKn9Vz9IFMC_SpCog6uP3-gFgY/view?usp=drive_link 

# from Webscraping
alpha_dir = 'data/data_prep/alpha_data/{}.csv' #.format(ticker) # I saved the raw alphadvantage data, so I don't have to do it again
yahoo_data = pd.read_csv('data/data_prep/yahoo_data.csv', index_col=0)
alpha_data = pd.read_csv('data/data_prep/alpha_data.csv', index_col=0)
MAEC_yahoo_data = pd.read_csv('data/data_prep/MAEC_yahoo_data.csv', index_col=0)
MAEC_alpha_data = pd.read_csv('data/data_prep/MAEC_alpha_data.csv', index_col=0)

yahoo_data.index = pd.to_datetime(yahoo_data.index)
alpha_data.index = pd.to_datetime(alpha_data.index)
MAEC_yahoo_data.index = pd.to_datetime(MAEC_yahoo_data.index)
MAEC_alpha_data.index = pd.to_datetime(MAEC_alpha_data.index)

# from Feature Engineering
glove_features = pd.read_csv('data/data_prep/glove_features.csv')
praat_features = pd.read_csv('data/data_prep/praat_features.csv', low_memory=False)
MAEC_glove_features = pd.read_csv('data/data_prep/MAEC_glove_features.csv')
MAEC_praat_features = pd.read_csv('data/data_prep/MAEC_praat_features.csv', low_memory=False)

# from Feature Engineering MORE
RoBERTa_features = pd.read_csv('data/data_prep/RoBERTa_features.csv', low_memory=False)
MAEC_RoBERTa_features = pd.read_csv('data/data_prep/MAEC_RoBERTa_features.csv', low_memory=False)
# Roberta with averages
RoBERTa_features2 = pd.read_csv('data/data_prep/RoBERTa_features2.csv', low_memory=False)
MAEC_RoBERTa_features2 = pd.read_csv('data/data_prep/MAEC_RoBERTa_features2.csv', low_memory=False)

investopedia_features = pd.read_csv('data/data_prep/investopedia_features.csv', low_memory=False)
MAEC_investopedia_features = pd.read_csv('data/data_prep/MAEC_investopedia_features.csv', low_memory=False)
bge_features = pd.read_csv('data/data_prep/bge_features.csv', low_memory=False)
MAEC_bge_features = pd.read_csv('data/data_prep/MAEC_bge_features.csv', low_memory=False)
bge_base_features = pd.read_csv('data/data_prep/bge_base_features.csv', low_memory=False)
MAEC_bge_base_features = pd.read_csv('data/data_prep/MAEC_bge_base_features.csv', low_memory=False)

# Loop through the directory, each folder represents an earnings conference call; the folders are named as "CompanyName_Date".
filename_data = []
for filename in os.listdir(original_data_dir):
    company_name, date_str = filename.rsplit('_', 1)
    date_str = date_str.split('.')[0] 
    date = datetime.strptime(date_str, "%Y%m%d").strftime("%Y-%m-%d")
    filename_data.append([company_name, date])
filename_data = pd.DataFrame(filename_data, columns=["Company", "Date"])
company_ticker = pd.read_csv('data/data_prep/company_ticker.csv')
filename_data = filename_data.merge(company_ticker, on="Company", how="left")

# Loop through the directory, each folder represents an earnings conference call; the folders are named as "Date_CompanyName".
MAEC_filename_data = []
for filename in os.listdir(MAEC_dir):
    date_str, ticker = filename.rsplit('_', 1)
    date_str = date_str.split('.')[0] 
    date = datetime.strptime(date_str, "%Y%m%d").strftime("%Y-%m-%d")
    MAEC_filename_data.append([ticker, date])
MAEC_filename_data = pd.DataFrame(MAEC_filename_data, columns=["Ticker", "Date"])

# Add TARGET of the regression

**n-day volatility predictions**: The predicted average volatility over the following n days.<br>

$$
v[0,n] = \ln \left( \sqrt{ \frac{1}{n} \sum_{i=1}^{n} (r_i - \bar{r})^2 } \right)
$$

Where:
- \( r_i \) is the stock return on day \(i\),
- \( \bar{r} \) is the average stock return over \(n\) days.

The stock return \(r_i\) is defined as:

$$
r_i = \frac{P_i - P_{i-1}}{P_{i-1}}
$$

Where \(P_i\) is the adjusted closing price of the stock on day \(i\).

For **single-day log volatility**, we estimate it using the **daily log absolute return**:

$$
v_n = \ln \left( \left| \frac{P_n - P_{n-1}}{P_{n-1}} \right| \right)
$$

Where:
- \(P_n\) is the adjusted closing price of the stock on day \(n\),
- \(P_{n-1}\) is the adjusted closing price on the previous day.

Our multi-task learning objective is to **simultaneously predict** these two quantities:
- \(v[0,n]\): The average volatility over \(n\) days (the main task).
- \(v_n\): The single-day volatility (the auxiliary task).


targets = filename_data.copy()

# Yahoo (missing companies)
def add_n_day(row, n_day):
    Ticker = row['Ticker']
    if Ticker in ['GGP', 'CA', 'STI', 'FLT', 'NLSN', 'WRK','RTN', 'UTX', 'DISH', 'VIACOM', 
                     'TIF', 'CELG', 'RHT', 'FLIR', 'AGN', 'KSU', 'NBL', 'ALXN', 'ABMD ', 'CTXS', 
                     'CBS', 'XL', 'VAR', 'ATVI', 'XLNX', 'SIVB']:
        return float('inf'), float('inf')
    Date = pd.to_datetime(row['Date'])
    Date_index = yahoo_data.index.get_loc(Date)
    data = yahoo_data.iloc[Date_index:(Date_index + 1 + n_day)][f"{Ticker}_Adj Close"]  
    #data = yahoo_data.iloc[(Date_index-1):(Date_index + n_day)][f"{Ticker}_Adj Close"]  
    stock_return = data.pct_change(fill_method=None).dropna() # ri =(Pi − Pi−1)/Pi−1
    std_dev = np.std(stock_return)
    if pd.isna(std_dev) or std_dev == 0:
        return 0, 0
    else:
        return np.log(std_dev), np.log(abs(stock_return.iloc[-1]) + 1e-10) #log 0 is undefined

targets['3_day'], targets['3_day_single'] = zip(*targets.progress_apply(lambda row: add_n_day(row, 3), axis=1))
targets['7_day'], targets['7_day_single'] = zip(*targets.progress_apply(lambda row: add_n_day(row, 7), axis=1))
targets['15_day'], targets['15_day_single'] = zip(*targets.progress_apply(lambda row: add_n_day(row, 15), axis=1))
targets['30_day'], targets['30_day_single'] = zip(*targets.progress_apply(lambda row: add_n_day(row, 30), axis=1))
targets.info(verbose=True)

# using alphadvantage 
def add_n_day(row, n_day):
    Ticker = row['Ticker'].strip()
    if Ticker in ['FLT', 'VIACOM', 'BA', 'BIIB', 'BDX', 'ABMD', 'CBS']:
        return 0, 0
    Date = pd.to_datetime(row['Date'])
    Date_index = alpha_data.index.get_loc(Date)
    data = alpha_data.iloc[Date_index:(Date_index + 1 + n_day)][Ticker]  
    #data = alpha_data.iloc[(Date_index-1):(Date_index + n_day)][Ticker]  
    stock_return = data.pct_change(fill_method=None).dropna() # ri =(Pi − Pi−1)/Pi−1
    std_dev = np.std(stock_return)
    if pd.isna(std_dev) or std_dev == 0:
        return 0, 0
    else:
        return np.log(std_dev), np.log(abs(stock_return.iloc[-1]) + 1e-10)

targets['3_day_alpha'], targets['3_day_single_alpha'] = zip(*targets.progress_apply(lambda row: add_n_day(row, 3), axis=1))
targets['7_day_alpha'], targets['7_day_single_alpha'] = zip(*targets.progress_apply(lambda row: add_n_day(row, 7), axis=1))
targets['15_day_alpha'], targets['15_day_single_alpha'] = zip(*targets.progress_apply(lambda row: add_n_day(row, 15), axis=1))
targets['30_day_alpha'], targets['30_day_single_alpha'] = zip(*targets.progress_apply(lambda row: add_n_day(row, 30), axis=1))

Here I am comparing the Yahoo data to alphadvantage. Why is there a difference? How does each calculate dividend adjustment?

targets['3_day_diff'] = targets['3_day_alpha'] - targets['3_day']
targets['7_day_diff'] = targets['7_day_alpha'] - targets['7_day']
targets['15_day_diff'] = targets['15_day_alpha'] - targets['15_day']
targets['30_day_diff'] = targets['30_day_alpha'] - targets['30_day']

targets['3_day_pct_change'] = abs(((targets['3_day_alpha'] - targets['3_day']) / targets['3_day_alpha']) * 100)
targets['7_day_pct_change'] = abs(((targets['7_day_alpha'] - targets['7_day']) / targets['7_day_alpha']) * 100)
targets['15_day_pct_change'] = abs(((targets['15_day_alpha'] - targets['15_day']) / targets['15_day_alpha']) * 100)
targets['30_day_pct_change'] = abs(((targets['30_day_alpha'] - targets['30_day']) / targets['30_day_alpha']) * 100)
 
targets['3_day_single_diff'] = targets['3_day_single_alpha'] - targets['3_day_single']
targets['7_day_single_diff'] = targets['7_day_single_alpha'] - targets['7_day_single']
targets['15_day_single_diff'] = targets['15_day_single_alpha'] - targets['15_day_single']
targets['30_day_single_diff'] = targets['30_day_single_alpha'] - targets['30_day_single']

targets['3_day_single_pct_change'] = abs(((targets['3_day_single_alpha'] - targets['3_day_single']) / targets['3_day_single_alpha']) * 100)
targets['7_day_single_pct_change'] = abs(((targets['7_day_single_alpha'] - targets['7_day_single']) / targets['7_day_single_alpha']) * 100)
targets['15_day_single_pct_change'] = abs(((targets['15_day_single_alpha'] - targets['15_day_single']) / targets['15_day_single_alpha']) * 100)
targets['30_day_single_pct_change'] = abs(((targets['30_day_single_alpha'] - targets['30_day_single']) / targets['30_day_single_alpha']) * 100)

# investigate discrepancies
targets = targets.sort_values(by='30_day_pct_change', ascending=False)
targets = targets.sort_values(by='30_day_single_pct_change', ascending=False)
targets.to_csv('data/data_prep/temp.csv', index=False) 

# Change Yahoo errors to the values found with alphadvantage 
targets['3_day_single'] = np.where(targets['3_day'] == float('inf'), targets['3_day_single_alpha'], targets['3_day_single'])
targets['7_day_single'] = np.where(targets['7_day'] == float('inf'), targets['7_day_single_alpha'], targets['7_day_single'])
targets['15_day_single'] = np.where(targets['15_day'] == float('inf'), targets['15_day_single_alpha'], targets['15_day_single'])
targets['30_day_single'] = np.where(targets['30_day'] == float('inf'), targets['30_day_single_alpha'], targets['30_day_single'])
targets['3_day'] = np.where(targets['3_day'] == float('inf'), targets['3_day_alpha'], targets['3_day'])
targets['7_day'] = np.where(targets['7_day'] == float('inf'), targets['7_day_alpha'], targets['7_day'])
targets['15_day'] = np.where(targets['15_day'] == float('inf'), targets['15_day_alpha'], targets['15_day'])
targets['30_day'] = np.where(targets['30_day'] == float('inf'), targets['30_day_alpha'], targets['30_day'])

targets['3_day_single'] = np.where(targets['3_day'] == 0, targets['3_day_single_alpha'], targets['3_day_single'])
targets['7_day_single'] = np.where(targets['7_day'] == 0, targets['7_day_single_alpha'], targets['7_day_single'])
targets['15_day_single'] = np.where(targets['15_day'] == 0, targets['15_day_single_alpha'], targets['15_day_single'])
targets['30_day_single'] = np.where(targets['30_day'] == 0, targets['30_day_single_alpha'], targets['30_day_single'])
targets['3_day'] = np.where(targets['3_day'] == 0, targets['3_day_alpha'], targets['3_day'])
targets['7_day'] = np.where(targets['7_day'] == 0, targets['7_day_alpha'], targets['7_day'])
targets['15_day'] = np.where(targets['15_day'] == 0, targets['15_day_alpha'], targets['15_day'])
targets['30_day'] = np.where(targets['30_day'] == 0, targets['30_day_alpha'], targets['30_day'])

print('Number of rows without data-', len(targets[((targets['3_day'] == 0) & 
                                                    (targets['7_day'] == 0) & 
                                                    (targets['15_day'] == 0) & 
                                                    (targets['30_day'] == 0))]))
print('----------------------------------------------------------------------')

# delete?
# targets = targets[~((targets['3_day'] == 0) & 
#                 (targets['7_day'] == 0) & 
#                 (targets['15_day'] == 0) & 
#                 (targets['30_day'] == 0))]

targets = targets.drop(['3_day_alpha', '3_day_single_alpha', '7_day_alpha', '7_day_single_alpha', '15_day_alpha', 
                        '15_day_single_alpha', '30_day_alpha', '30_day_single_alpha', '3_day_diff', '7_day_diff', 
                        '15_day_diff', '30_day_diff', '3_day_pct_change', '7_day_pct_change', '15_day_pct_change', 
                        '30_day_pct_change', '3_day_single_diff', '7_day_single_diff', '15_day_single_diff', 
                        '30_day_single_diff', '3_day_single_pct_change', '7_day_single_pct_change', 
                        '15_day_single_pct_change', '30_day_single_pct_change'], axis=1)
targets = targets.sort_values(by='3_day', ascending=True)
targets.info(verbose=True)
### save ############################################
targets.to_csv('data/data_prep/targets.csv', index=False)
#####################################################

# MAEC dataset TARGETS

[paper](https://dl.acm.org/doi/10.1145/3340531.3412879)
[GitHub](https://github.com/Earnings-Call-Dataset/MAEC-A-Multimodal-Aligned-Earnings-Conference-Call-Dataset-for-Financial-Risk-Prediction/tree/master)

MAEC_targets = MAEC_filename_data.copy()

# Yahoo (missing 9 companies)
def add_n_day(row, n_day):
    Ticker = row['Ticker']
    Date = pd.to_datetime(row['Date'])
    Date_index = MAEC_yahoo_data.index.get_loc(Date)
    data = MAEC_yahoo_data.iloc[Date_index:(Date_index + 1 + n_day)][f"{Ticker}_Adj Close"]  
    # data = MAEC_yahoo_data.iloc[(Date_index-1):(Date_index + n_day)][f"{Ticker}_Adj Close"]  
    stock_return = data.pct_change(fill_method=None).dropna() # ri =(Pi − Pi−1)/Pi−1
    std_dev = np.std(stock_return)
    if pd.isna(std_dev) or std_dev == 0:
        return 0, 0
    else:
        return np.log(std_dev), np.log(abs(stock_return.iloc[-1]) + 1e-10)

MAEC_targets['3_day'], MAEC_targets['3_day_single'] = zip(*MAEC_targets.progress_apply(lambda row: add_n_day(row, 3), axis=1))
MAEC_targets['7_day'], MAEC_targets['7_day_single'] = zip(*MAEC_targets.progress_apply(lambda row: add_n_day(row, 7), axis=1))
MAEC_targets['15_day'], MAEC_targets['15_day_single'] = zip(*MAEC_targets.progress_apply(lambda row: add_n_day(row, 15), axis=1))
MAEC_targets['30_day'], MAEC_targets['30_day_single'] = zip(*MAEC_targets.progress_apply(lambda row: add_n_day(row, 30), axis=1))
MAEC_targets.info(verbose=True)

# 58 tickers that did not return data from alphadvantage
bad_tickers = ['GPS', 'JCP', 'TUP', 'BBT', 'MDP', 'LL', 'ABC', 'PKI', 'HFC', 'HSC', 'CBB', 'ILG', 'JCOM', 'EBIX', 'ENDP', 
               'BIG', 'ASNA', 'IVC', 'BCOR', 'INT', 'FRED', 'CAMP', 'ELY', 'COG', 'CLD', 'CRY', 'PLT', 'FRAN', 'ADS', 'CPSI', 
               'FBHS', 'CHFC', 'UIHC', 'OFC', 'TMST', 'FTD', 'SWM', 'WPG', 'WLTW', 'AKRX', 'BLL', 'DF', 'TLRD', 'SPN', 'CLI', 
               'ESV', 'RCII', 'ANTM', 'RE', 'NCR', 'NEWM', 'PEI', 'LCI', 'ERA', 'ACOR', 'FB', 'AAXN', 'NLS']
def add_n_day(row, n_day):
    Ticker = row['Ticker']
    if Ticker in bad_tickers:
        return float('inf'), float('inf')
    Date = pd.to_datetime(row['Date'])
    Date_index = MAEC_alpha_data.index.get_loc(Date)
    data = MAEC_alpha_data.iloc[Date_index:(Date_index + 1 + n_day)][Ticker]  
    # data = MAEC_alpha_data.iloc[(Date_index-1):(Date_index + n_day)][Ticker]  
    stock_return = data.pct_change(fill_method=None).dropna() # ri =(Pi − Pi−1)/Pi−1
    std_dev = np.std(stock_return)
    if pd.isna(std_dev) or std_dev == 0:
        return 0, 0
    else:
        return np.log(std_dev), np.log(abs(stock_return.iloc[-1]) + 1e-10)

MAEC_targets['3_day_alpha'], MAEC_targets['3_day_single_alpha'] = zip(*MAEC_targets.progress_apply(lambda row: add_n_day(row, 3), axis=1))
MAEC_targets['7_day_alpha'], MAEC_targets['7_day_single_alpha'] = zip(*MAEC_targets.progress_apply(lambda row: add_n_day(row, 7), axis=1))
MAEC_targets['15_day_alpha'], MAEC_targets['15_day_single_alpha'] = zip(*MAEC_targets.progress_apply(lambda row: add_n_day(row, 15), axis=1))
MAEC_targets['30_day_alpha'], MAEC_targets['30_day_single_alpha'] = zip(*MAEC_targets.progress_apply(lambda row: add_n_day(row, 30), axis=1))

MAEC_targets['3_day_diff'] = MAEC_targets['3_day_alpha'] - MAEC_targets['3_day']
MAEC_targets['7_day_diff'] = MAEC_targets['7_day_alpha'] - MAEC_targets['7_day']
MAEC_targets['15_day_diff'] = MAEC_targets['15_day_alpha'] - MAEC_targets['15_day']
MAEC_targets['30_day_diff'] = MAEC_targets['30_day_alpha'] - MAEC_targets['30_day']

MAEC_targets['3_day_pct_change'] = abs(((MAEC_targets['3_day_alpha'] - MAEC_targets['3_day']) / MAEC_targets['3_day_alpha']) * 100)
MAEC_targets['7_day_pct_change'] = abs(((MAEC_targets['7_day_alpha'] - MAEC_targets['7_day']) / MAEC_targets['7_day_alpha']) * 100)
MAEC_targets['15_day_pct_change'] = abs(((MAEC_targets['15_day_alpha'] - MAEC_targets['15_day']) / MAEC_targets['15_day_alpha']) * 100)
MAEC_targets['30_day_pct_change'] = abs(((MAEC_targets['30_day_alpha'] - MAEC_targets['30_day']) / MAEC_targets['30_day_alpha']) * 100)
 
MAEC_targets['3_day_single_diff'] = MAEC_targets['3_day_single_alpha'] - MAEC_targets['3_day_single']
MAEC_targets['7_day_single_diff'] = MAEC_targets['7_day_single_alpha'] - MAEC_targets['7_day_single']
MAEC_targets['15_day_single_diff'] = MAEC_targets['15_day_single_alpha'] - MAEC_targets['15_day_single']
MAEC_targets['30_day_single_diff'] = MAEC_targets['30_day_single_alpha'] - MAEC_targets['30_day_single']

MAEC_targets['3_day_single_pct_change'] = abs(((MAEC_targets['3_day_single_alpha'] - MAEC_targets['3_day_single']) / MAEC_targets['3_day_single_alpha']) * 100)
MAEC_targets['7_day_single_pct_change'] = abs(((MAEC_targets['7_day_single_alpha'] - MAEC_targets['7_day_single']) / MAEC_targets['7_day_single_alpha']) * 100)
MAEC_targets['15_day_single_pct_change'] = abs(((MAEC_targets['15_day_single_alpha'] - MAEC_targets['15_day_single']) / MAEC_targets['15_day_single_alpha']) * 100)
MAEC_targets['30_day_single_pct_change'] = abs(((MAEC_targets['30_day_single_alpha'] - MAEC_targets['30_day_single']) / MAEC_targets['30_day_single_alpha']) * 100)
# investigate discrepancies
MAEC_targets = MAEC_targets.sort_values(by='30_day_pct_change', ascending=False)
MAEC_targets.to_csv('data/data_prep/MAEC_temp.csv', index=False)

# Change Yahoo errors to the values found with alphadvantage 
MAEC_targets['3_day_single'] = np.where(MAEC_targets['3_day'] == float('inf'), MAEC_targets['3_day_single_alpha'], MAEC_targets['3_day_single'])
MAEC_targets['7_day_single'] = np.where(MAEC_targets['7_day'] == float('inf'), MAEC_targets['7_day_single_alpha'], MAEC_targets['7_day_single'])
MAEC_targets['15_day_single'] = np.where(MAEC_targets['15_day'] == float('inf'), MAEC_targets['15_day_single_alpha'], MAEC_targets['15_day_single'])
MAEC_targets['30_day_single'] = np.where(MAEC_targets['30_day'] == float('inf'), MAEC_targets['30_day_single_alpha'], MAEC_targets['30_day_single'])
MAEC_targets['3_day'] = np.where(MAEC_targets['3_day'] == float('inf'), MAEC_targets['3_day_alpha'], MAEC_targets['3_day'])
MAEC_targets['7_day'] = np.where(MAEC_targets['7_day'] == float('inf'), MAEC_targets['7_day_alpha'], MAEC_targets['7_day'])
MAEC_targets['15_day'] = np.where(MAEC_targets['15_day'] == float('inf'), MAEC_targets['15_day_alpha'], MAEC_targets['15_day'])
MAEC_targets['30_day'] = np.where(MAEC_targets['30_day'] == float('inf'), MAEC_targets['30_day_alpha'], MAEC_targets['30_day'])

MAEC_targets['3_day_single'] = np.where(MAEC_targets['3_day'] == 0, MAEC_targets['3_day_single_alpha'], MAEC_targets['3_day_single'])
MAEC_targets['7_day_single'] = np.where(MAEC_targets['7_day'] == 0, MAEC_targets['7_day_single_alpha'], MAEC_targets['7_day_single'])
MAEC_targets['15_day_single'] = np.where(MAEC_targets['15_day'] == 0, MAEC_targets['15_day_single_alpha'], MAEC_targets['15_day_single'])
MAEC_targets['30_day_single'] = np.where(MAEC_targets['30_day'] == 0, MAEC_targets['30_day_single_alpha'], MAEC_targets['30_day_single'])
MAEC_targets['3_day'] = np.where(MAEC_targets['3_day'] == 0, MAEC_targets['3_day_alpha'], MAEC_targets['3_day'])
MAEC_targets['7_day'] = np.where(MAEC_targets['7_day'] == 0, MAEC_targets['7_day_alpha'], MAEC_targets['7_day'])
MAEC_targets['15_day'] = np.where(MAEC_targets['15_day'] == 0, MAEC_targets['15_day_alpha'], MAEC_targets['15_day'])
MAEC_targets['30_day'] = np.where(MAEC_targets['30_day'] == 0, MAEC_targets['30_day_alpha'], MAEC_targets['30_day'])

MAEC_targets.replace([float('inf'), -float('inf')], 0, inplace=True)
print('Number of rows without data-', len(MAEC_targets[((MAEC_targets['3_day'] == 0) & 
                                                        (MAEC_targets['7_day'] == 0) & 
                                                        (MAEC_targets['15_day'] == 0)& 
                                                        (MAEC_targets['30_day'] == 0))]))
print('----------------------------------------------------------------------')

# delete?
# MAEC_targets = MAEC_targets[~((MAEC_targets['3_day'] == 0) & 
#                 (MAEC_targets['7_day'] == 0) & 
#                 (MAEC_targets['15_day'] == 0) & 
#                 (MAEC_targets['30_day'] == 0))]

MAEC_targets = MAEC_targets.drop(['3_day_alpha', '3_day_single_alpha', '7_day_alpha', '7_day_single_alpha', '15_day_alpha', 
                        '15_day_single_alpha', '30_day_alpha', '30_day_single_alpha', '3_day_diff', '7_day_diff', 
                        '15_day_diff', '30_day_diff', '3_day_pct_change', '7_day_pct_change', '15_day_pct_change', 
                        '30_day_pct_change', '3_day_single_diff', '7_day_single_diff', '15_day_single_diff', 
                        '30_day_single_diff', '3_day_single_pct_change', '7_day_single_pct_change', 
                        '15_day_single_pct_change', '30_day_single_pct_change'], axis=1)
MAEC_targets = MAEC_targets.sort_values(by='7_day', ascending=True)
MAEC_targets.info(verbose=True)
### save ############################################
MAEC_targets.to_csv('data/data_prep/MAEC_targets.csv', index=False)
#####################################################

# Clean Praat features

- Replace '--undefined--' data
- Convert to numeric float64
- Compare our Praat data (MAEC), which was calculated from MP3 files.
    - To the provided Praat data (MAEC) [GitHub](https://github.com/Earnings-Call-Dataset/MAEC-A-Multimodal-Aligned-Earnings-Conference-Call-Dataset-for-Financial-Risk-Prediction/tree/master)

praat_features = praat_features.replace(['--undefined--', '--undefined-', '--undefined-- ',  '--'], np.nan)

# Convert all columns into float64 except 
cols_to_convert = praat_features.columns.difference(['Company', 'Date', 'audio_file'])
praat_features[cols_to_convert] = praat_features[cols_to_convert].progress_apply(pd.to_numeric)
# impute median on all NULL (--undefined--)
praat_features[cols_to_convert] = praat_features[cols_to_convert].progress_apply(
    lambda col: col.fillna(col.median()) if pd.api.types.is_numeric_dtype(col) else col
)
praat_features.info(verbose=True)

MAEC_praat_features = MAEC_praat_features.replace(['--undefined--', '--undefined-', '--undefined-- ',  '--'], np.nan)

# Convert all columns into float64 except 
cols_to_convert = MAEC_praat_features.columns.difference(['Ticker', 'Date', 'audio_file'])
MAEC_praat_features[cols_to_convert] = MAEC_praat_features[cols_to_convert].progress_apply(pd.to_numeric)
# impute median on all NULL (--undefined--)
MAEC_praat_features[cols_to_convert] = MAEC_praat_features[cols_to_convert].progress_apply(
    lambda col: col.fillna(col.median()) if pd.api.types.is_numeric_dtype(col) else col
)
MAEC_praat_features.info(verbose=True)

Compare

MAEC_GitHub_features = pd.DataFrame()
def each_row(row):
    Ticker = row['Ticker']
    Date = row['Date'].replace('-', '') 

    features_df = pd.read_csv(f"data/MAEC/MAEC_Dataset/{Date}_{Ticker}/features.csv")
    features_df['Ticker'] = Ticker
    features_df['Date'] = Date
    features_df['Sentence_num'] = range(1, len(features_df) + 1)

    global MAEC_GitHub_features
    MAEC_GitHub_features = pd.concat([MAEC_GitHub_features, features_df], ignore_index=True)

MAEC_filename_data.progress_apply(each_row, axis=1)

MAEC_GitHub_features = MAEC_GitHub_features.replace(['--undefined--', '--undefined-', '--undefined-- ',  '--'], np.nan)
# Convert all columns into float64 except 'Ticker', 'Date', 'audio_file' 
cols_to_convert = MAEC_GitHub_features.columns.difference(['Ticker', 'Date', 'audio_file'])
MAEC_GitHub_features[cols_to_convert] = MAEC_GitHub_features[cols_to_convert].progress_apply(pd.to_numeric)
# impute median on all NULL (--undefined--)
MAEC_GitHub_features[cols_to_convert] = MAEC_GitHub_features[cols_to_convert].progress_apply(
    lambda col: col.fillna(col.median()) if pd.api.types.is_numeric_dtype(col) else col
)
MAEC_GitHub_features.info(verbose=True)

# Compare
These are very close. That's good. This gives assurance that the original dataset Pratt features are okay.

MAEC_praat_features

MAEC_GitHub_features

# Final datasets


max_sent_len = 523

# The maximum sentences per meeting is 523 
# Ensure each meeting has 523 sentences, add zeros to the end
def add_zero_padding(group):
    complete_index = pd.Index(np.arange(1, (max_sent_len +1)), name='Sentence_num')
    group = group.set_index('Sentence_num').reindex(complete_index).reset_index()
    group['Date'] = group['Date'].ffill().bfill()
    group['Company'] = group['Company'].ffill().bfill()
    group.fillna(0.0, inplace=True)
    return group

def combine_audio_text_targets(audio_features, text_features, targets, num_features, save_directory):
    # combine & add_zero_padding
    features = pd.merge(audio_features, text_features, how="left",on = ['Company','Date','Sentence_num'])
    features = features.drop(['Shimmer apq11','Audio Length','audio_file'], axis=1)
    features = features.groupby(['Company', 'Date'], group_keys=False).progress_apply(add_zero_padding).reset_index(drop=True)
    features.fillna(0, inplace=True)

    # match the targets to the features
    # (target values will be duplicated 523 times, for each meeting, but that's okay for now)
    # we need the date and company columns to sort
    targets['Date'] = targets['Date'].astype(str).str.replace('-', '').astype('Int64')
    original_dataset = pd.merge(features, targets, how="left",on = ['Company','Date'])
    rows_with_nulls = original_dataset[original_dataset.isnull().any(axis=1)]
    original_dataset.fillna(0, inplace=True)
    print('Number of rows with NULL values -   ', len(rows_with_nulls))

    # dump rows without targets
    original_dataset = original_dataset.drop(['Ticker'], axis=1)
    original_dataset = original_dataset[~((original_dataset['3_day'] == 0) & 
                                        (original_dataset['7_day'] == 0) & 
                                        (original_dataset['15_day'] == 0) & 
                                        (original_dataset['30_day'] == 0))]
    original_dataset = original_dataset.sort_values(by=['Date', 'Company', 'Sentence_num'], ascending=[True, True, True])
    original_dataset.info(verbose=False)

    # train/val/test split FEATURES ONLY
    original_dataset['Date'] = pd.to_datetime(original_dataset['Date'].astype(str), format='%Y%m%d')
    train_features = original_dataset[original_dataset['Date'] <= '2017-08-02']
    val_features = original_dataset[(original_dataset['Date'] >= '2017-08-03') & (original_dataset['Date'] < '2017-10-24')]
    test_features = original_dataset[original_dataset['Date'] >= '2017-10-24']
    train_features = train_features.drop(['Company','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()
    val_features = val_features.drop(['Company','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()
    test_features = test_features.drop(['Company','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()

    # now for the TARGETS - we need 4 sets of targets (3, 7, 15, 30) each with primary/secondary targets 
    targets = targets[~((targets['3_day'] == 0) & 
                        (targets['7_day'] == 0) & 
                        (targets['15_day'] == 0) & 
                        (targets['30_day'] == 0))] # dump rows without targets
    targets = targets.copy().sort_values(by=['Date', 'Company'], ascending=[True, True])
    targets['Date'] = pd.to_datetime(targets['Date'].astype(str), format='%Y%m%d')
    # train/val/test split
    train_targets = targets[targets['Date'] <= '2017-08-02']
    val_targets = targets[(targets['Date'] >= '2017-08-03') & (targets['Date'] < '2017-10-24')]
    test_targets = targets[targets['Date'] >= '2017-10-24']

    #############################  save  ######################################################################
    # FEATURES- Reshape the NumPy array to have dimensions ( # meetings, 523 sentences (with padding), num_features)
    np.save(f'data/{save_directory}/train_features.npy', train_features.reshape(int(len(train_features)/max_sent_len), max_sent_len, num_features))
    np.save(f'data/{save_directory}/val_features.npy', val_features.reshape(int(len(val_features)/max_sent_len), max_sent_len, num_features))
    np.save(f'data/{save_directory}/test_features.npy', test_features.reshape(int(len(test_features)/max_sent_len), max_sent_len, num_features))
    
    np.save(f'data/{save_directory}/train_targets_3.npy', train_targets['3_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_3.npy', val_targets['3_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_3.npy', test_targets['3_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_3.npy', train_targets['3_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_3.npy', val_targets['3_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_3.npy', test_targets['3_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_7.npy', train_targets['7_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_7.npy', val_targets['7_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_7.npy', test_targets['7_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_7.npy', train_targets['7_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_7.npy', val_targets['7_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_7.npy', test_targets['7_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_15.npy', train_targets['15_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_15.npy', val_targets['15_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_15.npy', test_targets['15_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_15.npy', train_targets['15_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_15.npy', val_targets['15_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_15.npy', test_targets['15_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_30.npy', train_targets['30_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_30.npy', val_targets['30_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_30.npy', test_targets['30_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_30.npy', train_targets['30_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_30.npy', val_targets['30_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_30.npy', test_targets['30_day_single'].to_numpy())

        

praat_features['Date'] = praat_features['Date'].astype('Int64')
combine_audio_text_targets(praat_features, RoBERTa_features, targets, 1051, 'RoBERTa')

combine_audio_text_targets(praat_features, RoBERTa_features2, targets, 1051, 'RoBERTa2')

combine_audio_text_targets(praat_features, investopedia_features, targets, 795, 'investopedia')

combine_audio_text_targets(praat_features, bge_features, targets, 1051, 'bge')

combine_audio_text_targets(praat_features, bge_base_features, targets, 795, 'bge_base')

combine_audio_text_targets(praat_features, glove_features, targets, 327, 'glove')

# Final MAEC dataset

max_sent_len = 500 

# The maximum sentences per meeting is  
# Ensure each meeting has  sentences, add zeros to the end
def add_zero_padding(group):
    complete_index = pd.Index(np.arange(1, (max_sent_len + 1)), name='Sentence_num')
    group = group.set_index('Sentence_num').reindex(complete_index).reset_index()
    group['Date'] = group['Date'].ffill().bfill()
    group['Ticker'] = group['Ticker'].ffill().bfill()
    group.fillna(0.0, inplace=True)
    return group

def combine_audio_text_targets(audio_features, text_features, targets, num_features, save_directory):
    # combine & add_zero_padding
    features = pd.merge(audio_features, text_features, how="left",on = ['Ticker','Date','Sentence_num'])
    features = features.drop(['Shimmer apq11','Audio Length','audio_file'], axis=1)
    features = features.groupby(['Ticker', 'Date'], group_keys=False).progress_apply(add_zero_padding).reset_index(drop=True)
    features.fillna(0, inplace=True)

    # match the targets to the features
    # (target values will be duplicated 500 times, for each meeting, but that's okay for now)
    # we need the date and company columns to sort
    targets['Date'] = targets['Date'].astype(str).str.replace('-', '').astype('Int64')
    MAEC_dataset = pd.merge(features, targets, how="left",on = ['Ticker','Date'])
    rows_with_nulls = MAEC_dataset[MAEC_dataset.isnull().any(axis=1)]
    MAEC_dataset.fillna(0, inplace=True)
    print('Number of rows with NULL values -   ', len(rows_with_nulls))

    # dump rows without targets
    MAEC_dataset = MAEC_dataset[~((MAEC_dataset['3_day'] == 0) & 
                                        (MAEC_dataset['7_day'] == 0) & 
                                        (MAEC_dataset['15_day'] == 0) & 
                                        (MAEC_dataset['30_day'] == 0))]
    MAEC_dataset = MAEC_dataset.sort_values(by=['Date', 'Ticker', 'Sentence_num'], ascending=[True, True, True])
    MAEC_dataset.info(verbose=False)

    # train/val/test split FEATURES ONLY
    MAEC_dataset['Date'] = pd.to_datetime(MAEC_dataset['Date'].astype(str), format='%Y%m%d')
    train_features_2015 = MAEC_dataset[MAEC_dataset['Date'] <= '2015-10-21']
    val_features_2015 = MAEC_dataset[(MAEC_dataset['Date'] >= '2015-10-22') & (MAEC_dataset['Date'] < '2015-10-28')]
    test_features_2015 = MAEC_dataset[(MAEC_dataset['Date'] >= '2015-10-29') & (MAEC_dataset['Date'] < '2016-01-01')]
    # 2016
    train_features_2016 = MAEC_dataset[(MAEC_dataset['Date'] >= '2016-01-01') & (MAEC_dataset['Date'] < '2016-08-03')]
    val_features_2016 = MAEC_dataset[(MAEC_dataset['Date'] >= '2016-08-03') & (MAEC_dataset['Date'] < '2016-08-12')]
    test_features_2016 = MAEC_dataset[(MAEC_dataset['Date'] >= '2016-08-12') & (MAEC_dataset['Date'] < '2017-01-01')]
    # 2017
    train_features_2017 = MAEC_dataset[(MAEC_dataset['Date'] >= '2017-01-01') & (MAEC_dataset['Date'] < '2017-11-07')]
    val_features_2017 = MAEC_dataset[(MAEC_dataset['Date'] >= '2017-11-07') & (MAEC_dataset['Date'] < '2018-02-12')]
    test_features_2017 = MAEC_dataset[(MAEC_dataset['Date'] >= '2018-02-15') & (MAEC_dataset['Date'] < '2019-01-01')]

    train_features_2015 = train_features_2015.drop(['Ticker','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()
    val_features_2015 = val_features_2015.drop(['Ticker','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()
    test_features_2015 = test_features_2015.drop(['Ticker','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()
    # 2016
    train_features_2016 = train_features_2016.drop(['Ticker','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()
    val_features_2016 = val_features_2016.drop(['Ticker','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()
    test_features_2016 = test_features_2016.drop(['Ticker','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()
    # 2017
    train_features_2017 = train_features_2017.drop(['Ticker','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()
    val_features_2017 = val_features_2017.drop(['Ticker','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()
    test_features_2017 = test_features_2017.drop(['Ticker','Date','Sentence_num','3_day','3_day_single','7_day','7_day_single', '15_day','15_day_single','30_day','30_day_single'], axis=1).to_numpy()

    # now for the TARGETS - we need 4 sets of targets (3, 7, 15, 30) each with primary/secondary targets 
    targets = targets[~((targets['3_day'] == 0) & 
                        (targets['7_day'] == 0) & 
                        (targets['15_day'] == 0) & 
                        (targets['30_day'] == 0))] # dump rows without targets
    targets = targets.copy().sort_values(by=['Date', 'Ticker'], ascending=[True, True])
    targets['Date'] = pd.to_datetime(targets['Date'].astype(str), format='%Y%m%d')
    # train/val/test split
    train_targets_2015 = targets[targets['Date'] <= '2015-10-21']
    val_targets_2015 = targets[(targets['Date'] >= '2015-10-22') & (targets['Date'] < '2015-10-28')]
    test_targets_2015 = targets[(targets['Date'] >= '2015-10-29') & (targets['Date'] < '2016-01-01')]
    # 2016
    train_targets_2016 = targets[(targets['Date'] >= '2016-01-01') & (targets['Date'] < '2016-08-03')]
    val_targets_2016 = targets[(targets['Date'] >= '2016-08-03') & (targets['Date'] < '2016-08-12')]
    test_targets_2016 = targets[(targets['Date'] >= '2016-08-12') & (targets['Date'] < '2017-01-01')]
    # 2017
    train_targets_2017 = targets[(targets['Date'] >= '2017-01-01') & (targets['Date'] < '2017-11-07')]
    val_targets_2017 = targets[(targets['Date'] >= '2017-11-07') & (targets['Date'] < '2018-02-12')]
    test_targets_2017 = targets[(targets['Date'] >= '2018-02-15') & (targets['Date'] < '2019-01-01')]

    #############################  save  ######################################################################
    # FEATURES- Reshape the NumPy array to have dimensions ( # meetings, # sentences (with padding), num_features)
    np.save(f'data/{save_directory}/train_features_2015.npy', train_features_2015.reshape(int((len(train_features_2015))/max_sent_len), max_sent_len, num_features))
    np.save(f'data/{save_directory}/val_features_2015.npy', val_features_2015.reshape(int((len(val_features_2015))/max_sent_len), max_sent_len, num_features))
    np.save(f'data/{save_directory}/test_features_2015.npy', test_features_2015.reshape(int((len(test_features_2015))/max_sent_len), max_sent_len, num_features))
    # 2016
    np.save(f'data/{save_directory}/train_features_2016.npy', train_features_2016.reshape(int((len(train_features_2016))/max_sent_len), max_sent_len, num_features))
    np.save(f'data/{save_directory}/val_features_2016.npy', val_features_2016.reshape(int((len(val_features_2016))/max_sent_len), max_sent_len, num_features))
    np.save(f'data/{save_directory}/test_features_2016.npy', test_features_2016.reshape(int((len(test_features_2016))/max_sent_len), max_sent_len, num_features))
    # 2017
    np.save(f'data/{save_directory}/train_features_2017.npy', train_features_2017.reshape(int((len(train_features_2017))/max_sent_len), max_sent_len, num_features))
    np.save(f'data/{save_directory}/val_features_2017.npy', val_features_2017.reshape(int((len(val_features_2017))/max_sent_len), max_sent_len, num_features))
    np.save(f'data/{save_directory}/test_features_2017.npy', test_features_2017.reshape(int((len(test_features_2017))/max_sent_len), max_sent_len, num_features))

    ##################################################################################################
    # 2015 targets
    np.save(f'data/{save_directory}/train_targets_2015_3.npy', train_targets_2015['3_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2015_3.npy', val_targets_2015['3_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2015_3.npy', test_targets_2015['3_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2015_3.npy', train_targets_2015['3_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2015_3.npy', val_targets_2015['3_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2015_3.npy', test_targets_2015['3_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_2015_7.npy', train_targets_2015['7_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2015_7.npy', val_targets_2015['7_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2015_7.npy', test_targets_2015['7_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2015_7.npy', train_targets_2015['7_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2015_7.npy', val_targets_2015['7_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2015_7.npy', test_targets_2015['7_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_2015_15.npy', train_targets_2015['15_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2015_15.npy', val_targets_2015['15_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2015_15.npy', test_targets_2015['15_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2015_15.npy', train_targets_2015['15_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2015_15.npy', val_targets_2015['15_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2015_15.npy', test_targets_2015['15_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_2015_30.npy', train_targets_2015['30_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2015_30.npy', val_targets_2015['30_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2015_30.npy', test_targets_2015['30_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2015_30.npy', train_targets_2015['30_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2015_30.npy', val_targets_2015['30_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2015_30.npy', test_targets_2015['30_day_single'].to_numpy())

    ##################################################################################################
    # 2016 targets
    np.save(f'data/{save_directory}/train_targets_2016_3.npy', train_targets_2016['3_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2016_3.npy', val_targets_2016['3_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2016_3.npy', test_targets_2016['3_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2016_3.npy', train_targets_2016['3_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2016_3.npy', val_targets_2016['3_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2016_3.npy', test_targets_2016['3_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_2016_7.npy', train_targets_2016['7_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2016_7.npy', val_targets_2016['7_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2016_7.npy', test_targets_2016['7_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2016_7.npy', train_targets_2016['7_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2016_7.npy', val_targets_2016['7_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2016_7.npy', test_targets_2016['7_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_2016_15.npy', train_targets_2016['15_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2016_15.npy', val_targets_2016['15_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2016_15.npy', test_targets_2016['15_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2016_15.npy', train_targets_2016['15_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2016_15.npy', val_targets_2016['15_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2016_15.npy', test_targets_2016['15_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_2016_30.npy', train_targets_2016['30_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2016_30.npy', val_targets_2016['30_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2016_30.npy', test_targets_2016['30_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2016_30.npy', train_targets_2016['30_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2016_30.npy', val_targets_2016['30_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2016_30.npy', test_targets_2016['30_day_single'].to_numpy())

    ##################################################################################################
    # 2017
    np.save(f'data/{save_directory}/train_targets_2017_3.npy', train_targets_2017['3_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2017_3.npy', val_targets_2017['3_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2017_3.npy', test_targets_2017['3_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2017_3.npy', train_targets_2017['3_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2017_3.npy', val_targets_2017['3_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2017_3.npy', test_targets_2017['3_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_2017_7.npy', train_targets_2017['7_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2017_7.npy', val_targets_2017['7_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2017_7.npy', test_targets_2017['7_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2017_7.npy', train_targets_2017['7_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2017_7.npy', val_targets_2017['7_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2017_7.npy', test_targets_2017['7_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_2017_15.npy', train_targets_2017['15_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2017_15.npy', val_targets_2017['15_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2017_15.npy', test_targets_2017['15_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2017_15.npy', train_targets_2017['15_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2017_15.npy', val_targets_2017['15_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2017_15.npy', test_targets_2017['15_day_single'].to_numpy())

    np.save(f'data/{save_directory}/train_targets_2017_30.npy', train_targets_2017['30_day'].to_numpy())
    np.save(f'data/{save_directory}/val_targets_2017_30.npy', val_targets_2017['30_day'].to_numpy())
    np.save(f'data/{save_directory}/test_targets_2017_30.npy', test_targets_2017['30_day'].to_numpy())
    np.save(f'data/{save_directory}/train_secondary_targets_2017_30.npy', train_targets_2017['30_day_single'].to_numpy())
    np.save(f'data/{save_directory}/val_secondary_targets_2017_30.npy', val_targets_2017['30_day_single'].to_numpy())
    np.save(f'data/{save_directory}/test_secondary_targets_2017_30.npy', test_targets_2017['30_day_single'].to_numpy())




   

MAEC_praat_features['Date'] = MAEC_praat_features['Date'].astype('Int64')
combine_audio_text_targets(MAEC_praat_features, MAEC_RoBERTa_features, MAEC_targets, 1051, 'MAEC_RoBERTa')

combine_audio_text_targets(MAEC_praat_features, MAEC_RoBERTa_features2, MAEC_targets, 1051, 'MAEC_RoBERTa2')

combine_audio_text_targets(MAEC_praat_features, MAEC_investopedia_features, MAEC_targets, 795, 'MAEC_investopedia')

combine_audio_text_targets(MAEC_praat_features, MAEC_bge_features, MAEC_targets, 1051, 'MAEC_bge')

combine_audio_text_targets(MAEC_praat_features, MAEC_bge_base_features, MAEC_targets, 795, 'MAEC_bge_base')

combine_audio_text_targets(MAEC_praat_features, MAEC_glove_features, MAEC_targets, 327, 'MAEC_glove')
