==> 6-Results_Summary.ipynb <==
# Earnings Call Project: Data Cleaning
<br>
CIS 831 Deep Learning – Term Project<br>
Kansas State University
<br><br>
James Chapman<br>
John Woods<br>
Nathan Diehl<br>
<br>

import pandas as pd
import numpy as np
from tqdm import tqdm

final_results = pd.read_csv('data/KeFVP_final_results.csv')
final_results = final_results.drop(columns=['best_alpha'])
final_results = final_results.pivot(index='Embedding Variant', columns='n_days', values='MSE_testset')

KeFVP_MAEC15_final_results = pd.read_csv('data/KeFVP_MAEC15_final_results.csv')
KeFVP_MAEC15_final_results = KeFVP_MAEC15_final_results.drop(columns=['year','best_alpha'])
KeFVP_MAEC15_final_results = KeFVP_MAEC15_final_results.pivot(index='Embedding Variant', columns='n_days', values='MSE_testset')

KeFVP_MAEC16_final_results = pd.read_csv('data/KeFVP_MAEC16_final_results.csv')
KeFVP_MAEC16_final_results = KeFVP_MAEC16_final_results.drop(columns=['year','best_alpha'])
KeFVP_MAEC16_final_results = KeFVP_MAEC16_final_results.pivot(index='Embedding Variant', columns='n_days', values='MSE_testset')

# emotion2vec
KeFVP_emotion2vec_final_results = pd.read_csv('data/KeFVP_emotion2vec_final_results.csv')
KeFVP_emotion2vec_final_results = KeFVP_emotion2vec_final_results.drop(columns=['best_alpha'])
KeFVP_emotion2vec_final_results = KeFVP_emotion2vec_final_results.pivot(index='Embedding Variant', columns='n_days', values='MSE_testset')

KeFVP_emotion2vec_MAEC15_final_results = pd.read_csv('data/KeFVP_emotion2vec_MAEC15_final_results.csv')
KeFVP_emotion2vec_MAEC15_final_results = KeFVP_emotion2vec_MAEC15_final_results.drop(columns=['year','best_alpha'])
KeFVP_emotion2vec_MAEC15_final_results = KeFVP_emotion2vec_MAEC15_final_results.pivot(index='Embedding Variant', columns='n_days', values='MSE_testset')

KeFVP_emotion2vec_MAEC16_final_results = pd.read_csv('data/KeFVP_emotion2vec_MAEC16_final_results.csv')
KeFVP_emotion2vec_MAEC16_final_results = KeFVP_emotion2vec_MAEC16_final_results.drop(columns=['year','best_alpha'])
KeFVP_emotion2vec_MAEC16_final_results = KeFVP_emotion2vec_MAEC16_final_results.pivot(index='Embedding Variant', columns='n_days', values='MSE_testset')


final_results

KeFVP_MAEC15_final_results

KeFVP_MAEC16_final_results

KeFVP_emotion2vec_final_results

KeFVP_emotion2vec_MAEC15_final_results

KeFVP_emotion2vec_MAEC16_final_results

:)
